UpdateBD_Class

+++ src/main/java/com/example/aino_1/entity/HoaDonChiTiet.java
@@ -5,6 +5,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -29,6 +31,6 @@ public class HoaDonChiTiet {
     private Integer soLuong;  // Số lượng sản phẩm
 
     @Column(name = "gia", nullable = false)
-    private Float gia;  // Giá của sản phẩm trong chi tiết hóa đơn
+    private BigDecimal gia;  // Giá của sản phẩm trong chi tiết hóa đơn
 
 }
