+++ src/main/java/com/example/aino_1/entity/KichThuocLapTop.java
@@ -5,6 +5,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -16,7 +18,7 @@ public class KichThuocLapTop {
     private Integer id;
 
     @Column(name = "kich_thuoc", nullable = false)
-    private Float kichThuoc;
+    private BigDecimal kichThuoc;
 
     @Column(name = "trang_thai")
     private Integer trangThai;  // Cột trạng thái, mặc định là 1
